refactor(data-model): retire the classic shallowMutableClone(); promote the deep-frozen-children variant in its place
With every caller now wanting the "mutate the top, then deep-freeze the whole"
shape, the classic `shallowMutableClone()` (mutable top, identity-shared
children) no longer has any clients in the workspace. Removes it and renames
the sibling `shallowMutableDeepFrozenClone()` to `shallowMutableClone()`,
reworking its doc to drop the now-obsolete contrast paragraph and point would-be
structural-sharing callers at `cloneForMutation()` instead.

Caller imports renamed in lockstep across data-model and runner. The
`shallowMutableDeepFrozenClone.test.ts` file is renamed to
`shallowMutableClone.test.ts` (the old, obsolete `shallowMutableClone.test.ts`
that pinned the classic semantics is removed).
