+++ packages/data-model/src/fabric-value.ts
@@ -28,7 +28,6 @@ export {
   cloneWithoutValueAtPath,
   cloneWithValueAtPath,
   shallowMutableClone,
-  shallowMutableDeepFrozenClone,
 } from "./value-clone.ts";
 
 import type {
